pyproject.toml
[project]
name = "mcp_doc_summarize_edit"
version = "0.1.0"
description = "MCP server for document summarization and word replacement"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.14.2",
    "uv>=0.9.11",
    "streamlit>=1.53.0",
    "openai",
    "anthropic",
    "prompt-toolkit>=3.0.51",
    "pydantic-settings",
]

[project.scripts]
mcp-doc-summary = "mcp_document_summary.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_document_summary"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
]
